+++ x/mint/client/cli/query.go
@@ -19,6 +19,7 @@ func GetQueryCmd() *cobra.Command {
 	cmd.AddCommand(
 		GetCmdQueryParams(),
 		GetCmdQueryEpochProvisions(),
+		GetCmdQueryInflation(),
 	)
 
 	return cmd
@@ -74,6 +75,35 @@ func GetCmdQueryEpochProvisions() *cobra.Command {
 			}
 
 			return clientCtx.PrintString(fmt.Sprintf("%s\n", res.EpochProvisions))
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdQueryInflation implements a command to return the current minting
+// inflation value.
+func GetCmdQueryInflation() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "inflation",
+		Short: "Query the current minting inflation value",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+			queryClient := types.NewQueryClient(clientCtx)
+
+			params := &types.QueryInflationRequest{}
+			res, err := queryClient.Inflation(context.Background(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintString(fmt.Sprintf("%s\n", res.Inflation))
 		},
 	}